requirement.txt
langchain
langchain-community
langchain-core
langchainhub
langgraph
langchain-groq
langchain-astradb
langchain-nvidia-ai-endpoints
langchain-openai
langchain-neo4j
langchain-pinecone
langchain-fireworks
langchain_huggingface
langchain_chroma
langsmith
cassio
crewai
ollama
faiss-cpu
tiktoken
pypdf
PyPDF2
sentence-transformers
chromadb
beautifulsoup4
youtube-transcript-api
streamlit
gradio
sympy
numpy
python-dotenv
streamlit
PyMuPDF
duckduckgo_search
pytube
urllib3
huggingface_hub
crewai_tools
yt_dlp